- refactored DTOs - extended GUI for Customer management - added EnumValuesProvider to display enums-selector - added account.xhtml and implemented link function for it

+++ src/main/java/de/jreichl/jpa/repository/SingleEntityRepository.java
@@ -36,7 +36,7 @@ public abstract class SingleEntityRepository<E extends SingleEntity> {
         em.remove(entity);
     }
     
-    public E findById(int id) {
+    public E findById(long id) {
         return em.find(getEntityClass(), id);
     }
